MockEnv first draft.
this needs to change a bit to account for running sub-tests.  I think I'll make a few changes:

change NewMockEnv() to return a cleanup function that produces one output: a pointer to the MockLog struct that has the values from this tests's execution.

make methods on mockEnvImpl{} panic if the finish function has already been called.

make MockLog a recursive structure that contains a map[string]*MockLog for child tests.

handle any of the methods that can result in goroutine cancelation.